refactor(battle): modularize battle functions
